requirements-full.txt
# Sıfırdan kurulum veya tam bağımlılık listesi
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
gradio>=4.44.0
tensorboard>=2.10.0
